Fix clippy warnings in transit-router, transit-router-wasm, transit-data
Mechanical cleanups only, no behaviour change: redundant field names and
same-type casts, `let _ =` bindings of unit values, `is_none()` +
`unwrap()` → `map_or`, `% x == 0` → `is_multiple_of`, `format!("{e}")`
→ passing the `String` through, iterating `values()` instead of
`iter()`, a collapsible `if` in the tests, and an explicit
`allow(too_many_arguments)` on the wasm `compute_profile` entry point
whose flat argument list is the JS ABI. `cargo clippy --all-targets` is
now clean for these three crates.

+++ transit-router/src/path_display.rs
@@ -49,7 +49,7 @@ pub fn dominant_route_color(data: &PreparedData, path: &Path) -> Option<String> 
         .iter()
         .filter(|s| s.kind == SegmentKind::Transit)
         .max_by_key(|s| s.end_time.saturating_sub(s.start_time))?;
-    route_color(data, dominant.route_index? as u32)
+    route_color(data, dominant.route_index?)
 }
 
 fn format_segment(seg: &PathSegment) -> Vec<String> {
